+++ CHANGES.md
@@ -1,7 +1,23 @@
 CHANGES
 ========
 
-## 2.4.4-RC
+## 3.0.0-RC
+
+* Feature: Group Access Control List
+* Feature: Add theme selector
+* Feature: Custom title tag content
+* Feature: Add a control to switch whether email shown or hidden by user
+* Fix: bosai version
+* Support: Rename to GROWI
+* Support: Add dark theme
+* Support: Refreshing bootstrap theme and icons
+* Support: Use Browsersync instead of easy-livereload
+* Support: Upgrade libs
+    * react-bootstrap
+    * react-bootstrap-typeahead
+    * react-clipboard.js
+
+## 2.4.4
 
 * Feature: Autoformat Markdown Table
 * Feature: highlight.js Theme Selector